Set publisher and  rename README.md to DEVELOPMENT.md
The published extension uses the content of README.md, so the
original content suitable for development wasn't appropriate for
the published extension.  This commit renames README.md to
DEVELOPMENT.md and updates README.md to be suitable with a basic
description of the extension and a link to the published extension.